Move tests from SkijaInjectSample to skiko
