+++ samples/SkijaInjectSample/build.gradle.kts
@@ -69,17 +69,6 @@ tasks.register("runSoftware") {
     dependsOn(casualRun)
 }
 
-tasks.withType<Test> {
-    systemProperty("skiko.test.screenshots.dir", File(project.projectDir, "src/test/screenshots").absolutePath)
-
-    // Tests should be deterministic, so disable scaling.
-    // On MacOs we need the actual scale, otherwise we will have aliased screenshots because of scaling.
-    if (System.getProperty("os.name") != "Mac OS X") {
-        systemProperty("sun.java2d.dpiaware", "false")
-        systemProperty("sun.java2d.uiScale", "1")
-    }
-}
-
 tasks.withType<KotlinCompile>().configureEach {
     kotlinOptions.freeCompilerArgs += "-Xopt-in=kotlin.RequiresOptIn"
 }
