+++ skiko/src/jvmTest/kotlin/org/jetbrains/skiko/SkiaWindowTest.kt
@@ -1,11 +1,7 @@
 package org.jetbrains.skiko
 
 import kotlinx.coroutines.CompletableDeferred
-import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.delay
-import kotlinx.coroutines.runBlocking
-import kotlinx.coroutines.swing.Swing
 import org.jetbrains.skija.Canvas
 import org.jetbrains.skija.FontMgr
 import org.jetbrains.skija.Paint
@@ -14,6 +10,8 @@ import org.jetbrains.skija.paragraph.FontCollection
 import org.jetbrains.skija.paragraph.ParagraphBuilder
 import org.jetbrains.skija.paragraph.ParagraphStyle
 import org.jetbrains.skija.paragraph.TextStyle
+import org.jetbrains.skiko.util.ScreenshotTestRule
+import org.jetbrains.skiko.util.swingTest
 import org.junit.Assume.assumeTrue
 import org.junit.Rule
 import org.junit.Test
@@ -310,12 +308,6 @@ class SkiaWindowTest {
             screenshots.assert(window.bounds)
         } finally {
             window.close()
-        }
-    }
-
-    private fun swingTest(block: suspend CoroutineScope.() -> Unit) {
-        runBlocking(Dispatchers.Swing) {
-            block()
         }
     }
