+++ skiko/src/jvmTest/screenshots/macos/org_jetbrains_skiko_SkiaWindowTest_render_three_windows_window3.png
