+++ skiko/src/jvmTest/screenshots/windows_linux/org_jetbrains_skiko_SkiaWindowTest_render_single_window_frame1.png
