+++ skiko/src/jvmTest/screenshots/windows_linux/org_jetbrains_skiko_SkiaWindowTest_render_text_Windows.png
