+++ skiko/src/jvmTest/screenshots/windows_linux/org_jetbrains_skiko_SkiaWindowTest_resize_window.png
